@@ -5,9 +5,8 @@ import com.maxmind.geoip2.model.CityResponse;
 import com.maxmind.geoip2.record.*;
 import dev.vality.columbus.InvalidRequest;
 import org.apache.thrift.TException;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -15,6 +14,8 @@ import org.mockito.MockitoAnnotations;
 import java.io.IOException;
 
 import static dev.vality.columbus.columbusConstants.UNKNOWN;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 
 public class GeoIpServiceHandlerTest {
@@ -29,7 +30,7 @@ public class GeoIpServiceHandlerTest {
     @Mock
     Country country;
 
-    @Before
+    @BeforeEach
     public void init() throws IOException, GeoIp2Exception {
         MockitoAnnotations.initMocks(this);
         Mockito.when(country.getIsoCode()).thenReturn(RU);
@@ -42,24 +43,24 @@ public class GeoIpServiceHandlerTest {
     @Test
     public void getLocationIsoCode() throws TException {
         String locationIsoCode = geoIpServiceHandler.getLocationIsoCode(IP);
-        Assert.assertEquals(RUS, locationIsoCode);
+        assertEquals(RUS, locationIsoCode);
     }
 
     @Test
     public void getLocationIsoCodeUnknown() throws TException, IOException, GeoIp2Exception {
         Mockito.when(service.getLocationByIp(any())).thenReturn(null);
         String locationIsoCode = geoIpServiceHandler.getLocationIsoCode(IP);
-        Assert.assertEquals(UNKNOWN, locationIsoCode);
+        assertEquals(UNKNOWN, locationIsoCode);
     }
 
-    @Test(expected = InvalidRequest.class)
-    public void getLocationIsoCodeInvalidRequestTest() throws TException {
-        geoIpServiceHandler.getLocationIsoCode("23e23e23e2");
+    @Test
+    public void getLocationIsoCodeInvalidRequestTest() {
+        assertThrows(InvalidRequest.class, () -> geoIpServiceHandler.getLocationIsoCode("23e23e23e2"));
     }
 
-    @Test(expected = TException.class)
-    public void getLocationIsoCodeTExceptionTest() throws TException, IOException, GeoIp2Exception {
+    @Test
+    public void getLocationIsoCodeTExceptionTest() throws IOException, GeoIp2Exception {
         Mockito.when(service.getLocationByIp(any())).thenThrow(new RuntimeException());
-        geoIpServiceHandler.getLocationIsoCode("23e23e23e2");
+        assertThrows(TException.class, () -> geoIpServiceHandler.getLocationIsoCode("23e23e23e2"));
     }
 }
